requirements.txt
streamlit
requests
python-docx
